add a new CanonicalTileId extension method to calculate scale&offset of child/parent tiles change LatitudeElevationCompensation parameter to float from double fix map visualizer and composite layer module to support null coroutines on initial map load simplify vector layer visualizer constructor

+++ Runtime/Mapbox/BaseModule/Utilities/Conversions.cs
@@ -398,7 +398,7 @@ namespace Mapbox.BaseModule.Utilities
 			return angle * 6378.16;
 		}
 
-		public static float LatitudeElevationCompensation(float latitude)
+		public static float LatitudeElevationCompensation(double latitude)
 		{
 			return Mathf.Cos(Mathf.Deg2Rad * (float) latitude);
 		}
